Moved model to controller package. Added controller and service for vehicle.
