+++ src/main/java/pl/ciesielski/dominik/app/cardealerapp/controller/VehicleController.java
@@ -1,0 +1,4 @@
+package pl.ciesielski.dominik.app.cardealerapp.controller;
+
+public class VehicleController {
+}
