+++ src/main/java/pl/ciesielski/dominik/app/cardealerapp/dao/AddressDao.java
@@ -1,8 +1,8 @@
 package pl.ciesielski.dominik.app.cardealerapp.dao;
 
 import pl.ciesielski.dominik.app.cardealerapp.dao.utils.DatabaseConnectionManager;
-import pl.ciesielski.dominik.app.cardealerapp.model.Address;
-import pl.ciesielski.dominik.app.cardealerapp.model.AddressBuilder;
+import pl.ciesielski.dominik.app.cardealerapp.controller.model.Address;
+import pl.ciesielski.dominik.app.cardealerapp.controller.model.AddressBuilder;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
