+++ src/main/java/pl/ciesielski/dominik/app/cardealerapp/service/VehicleService.java
@@ -1,0 +1,4 @@
+package pl.ciesielski.dominik.app.cardealerapp.service;
+
+public class VehicleService {
+}
