+++ src/test/java/pl/ciesielski/dominik/app/cardealerapp/dao/ClientDaoIntegrationTest.java
@@ -8,9 +8,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.ciesielski.dominik.app.cardealerapp.dao.utils.DatabaseConnectionManager;
 import pl.ciesielski.dominik.app.cardealerapp.dao.utils.DatabaseInitializer;
-import pl.ciesielski.dominik.app.cardealerapp.model.Address;
-import pl.ciesielski.dominik.app.cardealerapp.model.AddressBuilder;
-import pl.ciesielski.dominik.app.cardealerapp.model.Client;
+import pl.ciesielski.dominik.app.cardealerapp.controller.model.Address;
+import pl.ciesielski.dominik.app.cardealerapp.controller.model.AddressBuilder;
+import pl.ciesielski.dominik.app.cardealerapp.controller.model.Client;
 import pl.ciesielski.dominik.app.cardealerapp.dao.utils.NextIdSequence;
 
 import java.sql.Connection;
